# Remove some allocation from Count and Measure