@mlerner 

```
benchmark                          old ns/op     new ns/op     delta
BenchmarkCountMultiple             749           386           -48.46%
BenchmarkCountMultipleWithRate     1786          684           -61.70%
BenchmarkMeasure                   754           379           -49.73%
BenchmarkMeasureWithRate           1782          687           -61.45%

benchmark                          old allocs     new allocs     delta
BenchmarkCountMultiple             5              1              -80.00%
BenchmarkCountMultipleWithRate     9              2              -77.78%
BenchmarkMeasure                   5              1              -80.00%
BenchmarkMeasureWithRate           9              2              -77.78%

benchmark                          old bytes     new bytes     delta
BenchmarkCountMultiple             139           101           -27.34%
BenchmarkCountMultipleWithRate     233           133           -42.92%
BenchmarkMeasure                   140           103           -26.43%
BenchmarkMeasureWithRate           235           135           -42.55%
```